+++ app/src/main/java/com/navapp/navigation/arrival/LackingPermissionException.java
@@ -1,4 +1,4 @@
-package com.navapp.navigation.util.exceptions;
+package com.navapp.navigation.arrival;
 
 public class LackingPermissionException extends Exception {
     public LackingPermissionException(String permissionName) {
