+++ app/src/main/java/com/navapp/navigation/arrival/SingleGeofenceInitializer.java
@@ -1,4 +1,4 @@
-package com.navapp.navigation.util.geofencing;
+package com.navapp.navigation.arrival;
 
 import android.Manifest;
 import android.app.PendingIntent;
@@ -14,7 +14,6 @@ import com.google.android.gms.location.GeofencingRequest;
 import com.google.android.gms.location.LocationServices;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.maps.model.LatLng;
-import com.navapp.navigation.util.exceptions.LackingPermissionException;
 
 import java.io.Closeable;
 import java.util.Collections;
